Add confirmation for deleting comments

+++ src/public/song/scripts/comments.js
@@ -31,10 +31,32 @@ const displayComments = {
 						try {
 							e.preventDefault();
 							e.stopPropagation();
-
-							await axios.delete(`${window.location.pathname}/comments/${comment._id}`);
-							newComment.remove();
+							const result = await Swal.fire({
+								cancelButtonColor  : "#d33",
+								confirmButtonColor : "#3085d6",
+								confirmButtonText  : "Confirm",
+								icon               : "warning",
+								showCancelButton   : true,
+								text               : "You won't be able to revert this!",
+								title              : "Delete your comment?",
+							});
+							if (result.isConfirmed) {
+								await axios.delete(`${window.location.pathname}/comments/${comment._id}`);
+								newComment.remove();
+								Swal.fire({
+									confirmButtonText : "Okay",
+									icon              : "success",
+									text              : "Your comment has been deleted!",
+									title             : "Success!"
+								});
+							}
 						} catch (err) {
+							Swal.fire({
+								confirmButtonText : "Okay",
+								icon              : "error",
+								text              : err.message,
+								title             : "Error!"
+							});
 							console.error(err);
 						}
 					});
